weatherbot_gcancio/requirements.txt
pyTelegramBotAPI
requests
python-dotenv 
python-decouple